refixed the download of excel file

+++ server/startup.js
@@ -7,31 +7,29 @@ Files = new FS.Collection("files", {
 
 // Wrapped everything in a setInterval function in order to update the info every now and then.
 // Set to run this all function every 2 days.
+// In any case, it will download a nw file only if the one we have is 1 month old.
 Meteor.startup(startup);
 Meteor.setInterval(startup, 2*24*60*60*1000);
 
-// Only run this code if a file has been stored.
-var excelfile = Files.findOne();
-if (excelfile && excelfile.hasStored("files")) {
-    
-    // Havving a file, we need to parse it in order to read it.
-    // made a function called parseExcel for that.
-    var workbook = parseExcel(excelfile);
-    console.log("parsed excel file.");
-    
-    // I know before hand that we need the 4th sheet.
-    var workSheet = workbook.Sheets[workbook.SheetNames[3]];
-    
-    // Finally, we need to insert the relevant data into the Mongo database.
-    // made a function called insertDB to deal with it.
-    var insertResult = insertDB(workSheet);
-    console.log(insertResult);
+var interval = Meteor.setInterval(parsing, 2000);
+
+function parsing(){
+    var excelfile = Files.findOne();
+    // Only run this code if a file has been stored.
+    if(excelfile && excelfile.hasStored("files")) {
+        var workbook = parseExcel(excelfile);
+        console.log("parsed excel file.");
+        
+        // I know before hand that we need the 4th sheet.
+        var workSheet = workbook.Sheets[workbook.SheetNames[3]];
+        
+        // Finally, we need to insert the relevant data into the Mongo database.
+        // made a function called insertDB to deal with it.
+        var insertResult = insertDB(workSheet);
+        console.log(insertResult);
+        Meteor.clearInterval(interval);
+    }
 }
-
-
-
-
-
 
 
 
@@ -40,7 +38,7 @@ if (excelfile && excelfile.hasStored("files")) {
 // This function is to retrieve the excel file from the given url
 //////////////////////////////////////
 function startup(){
-    console.log("Downloading the file...");
+    console.log("Testing if it needs to download the file...");
 
     var url = "http://www.dripinvesting.org/tools/U.S.DividendChampions.xls";
 
@@ -48,18 +46,25 @@ function startup(){
     // it was necessary to add the CollectionFS package.
     
     // First lets see if we have already a file stored
-    var objFile = Files.insert(url, function(err, file) {
-        if (err) {
-            console.log("Didn't inserted. Error", err);
-        } else {
-            // We don't need more than 1 downloaded file.
-            // So, upon sucess inserting new file, all others will be deleted.
-            Files.remove({_id: {$ne: file._id}});
-            console.log("Download sucessed. Inserted: " + Files.findOne()._id);
-        }
-    });
-    
-    return objFile;
+    var objFile = Files.findOne();
+
+    if (!objFile || objFile.updatedAt().getTime() < new Date() - 32*24*60*60*1000) {
+        
+        // if there is no file, or if the file is more than 1 month old, then insert/save a new file.
+        objFile = Files.insert(url, function(err, file) {
+            if (err) {
+                console.log("Didn't inserted. Error", err);
+            } else {
+                // We don't need more than 1 downloaded file.
+                // So, upon sucess inserting new file, all others will be deleted.
+                Files.remove({_id: {$ne: file._id}});
+                console.log("Download sucessed. Inserted: " + Files.findOne()._id);
+            }
+        });
+        return objFile;
+    } else {
+        console.log("And download was not necessary.");
+    }
 }
